requirements.txt
torchvision-cpu
torch-cpu
selenium
numpy
pandas

